CASS-48 | Tests refactor imports and move flows runners to specific tests classes
