+++ src/test/java/org/mule/modules/cassandradb/automation/functional/AbstractTestCases.java
@@ -70,245 +70,6 @@ public class AbstractTestCases extends MuleArtifactFunctionalTestCase {
         cassandraService.dropKeyspace(cassandraProperties.getKeyspace());
     }
 
-    boolean createKeyspace(final CreateKeyspaceInput keyspaceInput) throws Exception {
-        return (boolean) flowRunner("createKeyspace-flow")
-                .withPayload(keyspaceInput)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void createKeyspaceExpException(final CreateKeyspaceInput keyspaceInput, ErrorTypeDefinition errorType) throws Exception {
-        flowRunner("createKeyspace-flow")
-                .withPayload(keyspaceInput)
-                .runExpectingException(ErrorTypeMatcher.errorType(errorType));
-    }
-
-    boolean dropKeyspace(String keyspaceName) throws Exception {
-        return (boolean) flowRunner("deleteKeyspace-flow")
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    boolean createTable(CreateTableInput basicCreateTableInput) throws Exception {
-        return (boolean) flowRunner("createTable-flow")
-                .withPayload(basicCreateTableInput)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    boolean dropTable(String tableName, String keyspaceName) throws Exception {
-        return (boolean) flowRunner("dropTable-flow")
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    boolean addNewColumn(String tableName, String keyspaceName, AlterColumnInput alterColumnInput) throws Exception {
-        return (boolean) flowRunner("addColumn-flow")
-                .withPayload(alterColumnInput)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-     void addNewColumnExpException(String tableName, String keyspaceName, AlterColumnInput alterColumnInput) throws Exception {
-         flowRunner("addColumn-flow")
-                 .withPayload(alterColumnInput)
-                 .withVariable("tableName", tableName)
-                 .withVariable("keyspaceName", keyspaceName)
-                 .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    boolean dropColumn(String tableName, String keyspaceName, String column) throws Exception {
-        return (boolean) flowRunner("dropColumn-flow")
-                .withPayload(column)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void dropColumnExpException(String tableName, String keyspaceName, String column) throws Exception {
-        flowRunner("dropColumn-flow")
-                .withPayload(column)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    boolean renameColumn(String tableName, String keyspaceName, String column, String newColumn) throws Exception {
-        return (boolean) flowRunner("renameColumn-flow")
-                .withPayload(column)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .withVariable("newColumnName", newColumn)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void renameColumnExpException(String tableName, String keyspaceName, String column, String newColumn) throws Exception {
-        flowRunner("renameColumn-flow")
-                .withPayload(column)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .withVariable("newColumnName", newColumn)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    boolean changeColumnType(String tableName, String keyspaceName, AlterColumnInput input) throws Exception {
-        return (boolean) flowRunner("changeColumnType-flow")
-                .withPayload(input)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void changeColumnTypeExpException(String tableName, String keyspaceName, AlterColumnInput input) throws Exception {
-        flowRunner("changeColumnType-flow")
-                .withPayload(input)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    List<String> getTableNamesFromKeyspace(String keyspaceName) throws Exception {
-        return (List<String>) flowRunner("getTableNamesFromKeyspace-flow")
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void insert(String tableName, String keyspaceName, Map<String, Object> entity) throws Exception {
-        flowRunner("insert-flow")
-                .withPayload(entity)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    void insertExpError(String tableName, String keyspaceName, Map<String, Object> entity) throws Exception {
-        flowRunner("insert-flow")
-                .withPayload(entity)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    protected void update(String tableName, String keyspaceName, Map<String, Object> entityToUpdate) throws Exception {
-        flowRunner("update-flow")
-                .withPayload(entityToUpdate)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    protected void updateExpException(String tableName, String keyspaceName, Map<String, Object> entityToUpdate) throws Exception {
-        flowRunner("update-flow")
-                .withPayload(entityToUpdate)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    protected List<Map<String, Object>> select(String validParameterizedQuery, List<Object> validParmList) throws Exception {
-        return (List<Map<String, Object>>) flowRunner("select-flow")
-                .withPayload(validParameterizedQuery)
-                .withVariable("parameters", validParmList)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-
-    protected void selectExpException(String validParameterizedQuery, List<Object> validParmList) throws Exception {
-        flowRunner("select-flow")
-                .withPayload(validParameterizedQuery)
-                .withVariable("parameters", validParmList)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    protected List<Map<String, Object>> executeCQLQuery(CQLQueryInput query) throws Exception {
-        return (List<Map<String, Object>>) flowRunner("executeCQLQuery-flow")
-                .withPayload(query)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    protected void executeCQLQueryExpException(CQLQueryInput query) throws Exception {
-        flowRunner("executeCQLQuery-flow")
-                .withPayload(query)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    protected void deleteColumnsValue(String tableName, String keyspaceName, Map<String, Object> payloadColumnsAndFilters) throws Exception {
-        flowRunner("deleteColumnsValue-flow")
-                .withPayload(payloadColumnsAndFilters)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    protected void deleteColumnsValueExpException(String tableName, String keyspaceName, Map<String, Object> payloadColumnsAndFilters) throws Exception {
-        flowRunner("deleteColumnsValue-flow")
-                .withPayload(payloadColumnsAndFilters)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
-    protected void deleteRows(String tableName, String keyspaceName, Map<String, Object> payloadColumnsAndFilters) throws Exception {
-        flowRunner("deleteRows-flow")
-                .withPayload(payloadColumnsAndFilters)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .run()
-                .getMessage()
-                .getPayload()
-                .getValue();
-    }
-
-    protected void deleteRowsExpException(String tableName, String keyspaceName, Map<String, Object> payloadColumnsAndFilters) throws Exception {
-        flowRunner("deleteRows-flow")
-                .withPayload(payloadColumnsAndFilters)
-                .withVariable("tableName", tableName)
-                .withVariable("keyspaceName", keyspaceName)
-                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
-    }
-
     public static CassandraProperties getCassandraProperties() {
         //load required properties
         CassandraProperties cassProperties = null;
