+++ src/test/java/org/mule/modules/cassandradb/automation/functional/AddNewColumnTestCase.java
@@ -9,15 +9,20 @@ import org.mule.modules.cassandradb.api.AlterColumnInput;
 import org.mule.modules.cassandradb.api.ColumnType;
 import org.mule.modules.cassandradb.api.CreateTableInput;
 import org.mule.modules.cassandradb.automation.util.TestsConstants;
+import org.mule.modules.cassandradb.internal.exception.CassandraError;
+import org.mule.tck.junit4.matcher.ErrorTypeMatcher;
 
 import static org.junit.Assert.assertTrue;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getAlterColumnInput;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getBasicCreateTableInput;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getColumns;
 
 
 public class AddNewColumnTestCase extends AbstractTestCases {
 
     @Before
     public void setup() {
-        CreateTableInput basicCreateTableInput = TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), TestsConstants.TABLE_NAME_1);
+        CreateTableInput basicCreateTableInput = getBasicCreateTableInput(getColumns(), getKeyspaceFromProperties(), TestsConstants.TABLE_NAME_1);
         getCassandraService().createTable(basicCreateTableInput);
     }
 
@@ -28,13 +33,32 @@ public class AddNewColumnTestCase extends AbstractTestCases {
 
     @Test
     public void testAddNewColumnOfPrimitiveTypeWithSuccess() throws Exception {
-        AlterColumnInput alterColumnInput = TestDataBuilder.getAlterColumnInput(DataType.text().toString() + System.currentTimeMillis(), ColumnType.TEXT);
+        AlterColumnInput alterColumnInput = getAlterColumnInput(DataType.text().toString() + System.currentTimeMillis(), ColumnType.TEXT);
         assertTrue(addNewColumn(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties(), alterColumnInput));
     }
 
     @Test
     public void testAddNewColumnWithSameName() throws Exception {
-        AlterColumnInput alterColumnInput = TestDataBuilder.getAlterColumnInput(TestsConstants.VALID_COLUMN_1, ColumnType.TEXT);
+        AlterColumnInput alterColumnInput = getAlterColumnInput(TestsConstants.VALID_COLUMN_1, ColumnType.TEXT);
         addNewColumnExpException(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties(), alterColumnInput);
+    }
+
+    boolean addNewColumn(String tableName, String keyspaceName, AlterColumnInput alterColumnInput) throws Exception {
+        return (boolean) flowRunner("addColumn-flow")
+                .withPayload(alterColumnInput)
+                .withVariable("tableName", tableName)
+                .withVariable("keyspaceName", keyspaceName)
+                .run()
+                .getMessage()
+                .getPayload()
+                .getValue();
+    }
+
+    void addNewColumnExpException(String tableName, String keyspaceName, AlterColumnInput alterColumnInput) throws Exception {
+        flowRunner("addColumn-flow")
+                .withPayload(alterColumnInput)
+                .withVariable("tableName", tableName)
+                .withVariable("keyspaceName", keyspaceName)
+                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
     }
 }
