+++ src/test/java/org/mule/modules/cassandradb/automation/functional/DropTableTestCase.java
@@ -12,6 +12,8 @@ import org.mule.modules.cassandradb.automation.util.TestsConstants;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getBasicCreateTableInput;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getCompositePrimaryKey;
 
 public class DropTableTestCase extends AbstractTestCases {
 
@@ -19,7 +21,7 @@ public class DropTableTestCase extends AbstractTestCases {
     public void testDropTable() throws Exception {
         String tableName = TestsConstants.TABLE_NAME_1;
         String keyspace = getKeyspaceFromProperties();
-        CreateTableInput basicCreateTableInput = TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), keyspace, tableName);
+        CreateTableInput basicCreateTableInput = getBasicCreateTableInput(TestDataBuilder.getColumns(), keyspace, tableName);
         getCassandraService().createTable(basicCreateTableInput);
         assertNotNull(getTableMetadata(tableName, keyspace));
 
@@ -33,7 +35,7 @@ public class DropTableTestCase extends AbstractTestCases {
     public void testDropTableWithCompositePK() throws Exception {
         String tableName = TestsConstants.TABLE_NAME_2;
         String keyspace = getKeyspaceFromProperties();
-        CreateTableInput basicCreateTableInput = TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getCompositePrimaryKey(), keyspace, tableName);
+        CreateTableInput basicCreateTableInput = getBasicCreateTableInput(getCompositePrimaryKey(), keyspace, tableName);
         getCassandraService().createTable(basicCreateTableInput);
         assertNotNull(getTableMetadata(tableName, keyspace));
 
@@ -46,5 +48,15 @@ public class DropTableTestCase extends AbstractTestCases {
     private TableMetadata getTableMetadata(String tableName, String keyspaceName) {
         KeyspaceMetadata keyspaceMetadata = getKeyspaceMetadata(keyspaceName);
         return keyspaceMetadata.getTable(tableName);
+    }
+
+    boolean dropTable(String tableName, String keyspaceName) throws Exception {
+        return (boolean) flowRunner("dropTable-flow")
+                .withVariable("tableName", tableName)
+                .withVariable("keyspaceName", keyspaceName)
+                .run()
+                .getMessage()
+                .getPayload()
+                .getValue();
     }
 }
