+++ src/test/java/org/mule/modules/cassandradb/automation/functional/GetTableNamesFromKeyspaceTestCase.java
@@ -5,12 +5,12 @@ package org.mule.modules.cassandradb.automation.functional;
 
 import org.junit.Test;
 import org.mule.modules.cassandradb.api.CreateKeyspaceInput;
-import org.mule.modules.cassandradb.automation.util.TestsConstants;
 
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mule.modules.cassandradb.automation.functional.TestDataBuilder.getColumns;
 import static org.mule.modules.cassandradb.automation.util.TestsConstants.KEYSPACE_NAME_3;
 import static org.mule.modules.cassandradb.automation.util.TestsConstants.TABLE_NAME_1;
 import static org.mule.modules.cassandradb.automation.util.TestsConstants.TABLE_NAME_2;
@@ -19,8 +19,8 @@ public class GetTableNamesFromKeyspaceTestCase extends AbstractTestCases {
 
     @Test
     public void testTableNamesFromLoggedInKeyspace() throws Exception {
-        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), TABLE_NAME_1));
-        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), TABLE_NAME_2));
+        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(getColumns(), getKeyspaceFromProperties(), TABLE_NAME_1));
+        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(getColumns(), getKeyspaceFromProperties(), TABLE_NAME_2));
 
         List<String> tableNames = getTableNamesFromKeyspace(getKeyspaceFromProperties());
 
@@ -40,9 +40,9 @@ public class GetTableNamesFromKeyspaceTestCase extends AbstractTestCases {
 
         getCassandraService().createKeyspace(keyspaceInput);
 
-        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), KEYSPACE_NAME_3, TABLE_NAME_1));
-        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), KEYSPACE_NAME_3, TABLE_NAME_2));
-        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), testTable));
+        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(getColumns(), KEYSPACE_NAME_3, TABLE_NAME_1));
+        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(getColumns(), KEYSPACE_NAME_3, TABLE_NAME_2));
+        getCassandraService().createTable(TestDataBuilder.getBasicCreateTableInput(getColumns(), getKeyspaceFromProperties(), testTable));
 
         List<String> tableNames = getTableNamesFromKeyspace(KEYSPACE_NAME_3);
 
@@ -51,5 +51,14 @@ public class GetTableNamesFromKeyspaceTestCase extends AbstractTestCases {
         assertTrue(tableNames.contains(TABLE_NAME_2));
 
         getCassandraService().dropKeyspace(KEYSPACE_NAME_3);
+    }
+
+    List<String> getTableNamesFromKeyspace(String keyspaceName) throws Exception {
+        return (List<String>) flowRunner("getTableNamesFromKeyspace-flow")
+                .withVariable("keyspaceName", keyspaceName)
+                .run()
+                .getMessage()
+                .getPayload()
+                .getValue();
     }
 }
