+++ src/test/java/org/mule/modules/cassandradb/automation/functional/InsertTestCase.java
@@ -6,51 +6,77 @@ package org.mule.modules.cassandradb.automation.functional;
 import com.datastax.driver.core.DataType;
 import org.junit.*;
 import org.mule.modules.cassandradb.api.CreateTableInput;
-import org.mule.modules.cassandradb.automation.util.TestsConstants;
+import org.mule.modules.cassandradb.internal.exception.CassandraError;
+import org.mule.tck.junit4.matcher.ErrorTypeMatcher;
+
+import java.util.Map;
+
+import static org.mule.modules.cassandradb.automation.util.TestsConstants.TABLE_NAME_1;
+import static org.mule.modules.cassandradb.automation.util.TestsConstants.VALID_LIST_COLUMN;
+import static org.mule.modules.cassandradb.automation.util.TestsConstants.VALID_MAP_COLUMN;
+import static org.mule.modules.cassandradb.automation.util.TestsConstants.VALID_SET_COLUMN;
 
 public class InsertTestCase extends AbstractTestCases {
 
     @Before
     public void setup() throws Exception {
-        CreateTableInput basicCreateTableInput = TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), TestsConstants.TABLE_NAME_1);
+        CreateTableInput basicCreateTableInput = TestDataBuilder.getBasicCreateTableInput(TestDataBuilder.getColumns(), getKeyspaceFromProperties(), TABLE_NAME_1);
         getCassandraService().createTable(basicCreateTableInput);
     }
 
     @After
     public void tearDown()  {
-        getCassandraService().dropTable(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties());
+        getCassandraService().dropTable(TABLE_NAME_1, getKeyspaceFromProperties());
     }
 
     @Test
     public void testInsertWithSuccess() throws Exception {
-        insert(TestsConstants.TABLE_NAME_1, null, TestDataBuilder.getValidEntity());
+        insert(TABLE_NAME_1, null, TestDataBuilder.getValidEntity());
         //TODO: add validation that the objects are created.
     }
 
     @Test
     public void testInsertInListColumnWithSuccess() throws Exception {
-        getCassandraService().addNewColumn(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties(), TestsConstants.VALID_LIST_COLUMN, DataType.list(DataType.text()));
+        getCassandraService().addNewColumn(TABLE_NAME_1, getKeyspaceFromProperties(), VALID_LIST_COLUMN, DataType.list(DataType.text()));
 
-        insert(TestsConstants.TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithList());
+        insert(TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithList());
     }
 
     @Test
     public void testInsertInMapColumnWithSuccess() throws Exception {
-        getCassandraService().addNewColumn(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties(), TestsConstants.VALID_MAP_COLUMN, DataType.map(DataType.text(), DataType.text()));
+        getCassandraService().addNewColumn(TABLE_NAME_1, getKeyspaceFromProperties(), VALID_MAP_COLUMN, DataType.map(DataType.text(), DataType.text()));
 
-        insert(TestsConstants.TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithMap());
+        insert(TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithMap());
     }
 
     @Test
     public void testInsertISetColumnWithSuccess() throws Exception {
-        getCassandraService().addNewColumn(TestsConstants.TABLE_NAME_1, getKeyspaceFromProperties(), TestsConstants.VALID_SET_COLUMN, DataType.set(DataType.text()));
+        getCassandraService().addNewColumn(TABLE_NAME_1, getKeyspaceFromProperties(), VALID_SET_COLUMN, DataType.set(DataType.text()));
 
-        insert(TestsConstants.TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithSet());
+        insert(TABLE_NAME_1, null, TestDataBuilder.getValidEntityWithSet());
     }
 
     @Test
     public void testInsertWithInvalidInput() throws Exception {
-        insertExpError(TestsConstants.TABLE_NAME_1, null, TestDataBuilder.getInvalidEntity());
+        insertExpError(TABLE_NAME_1, null, TestDataBuilder.getInvalidEntity());
     }
 
+    void insert(String tableName, String keyspaceName, Map<String, Object> entity) throws Exception {
+        flowRunner("insert-flow")
+                .withPayload(entity)
+                .withVariable("tableName", tableName)
+                .withVariable("keyspaceName", keyspaceName)
+                .run()
+                .getMessage()
+                .getPayload()
+                .getValue();
+    }
+
+    void insertExpError(String tableName, String keyspaceName, Map<String, Object> entity) throws Exception {
+        flowRunner("insert-flow")
+                .withPayload(entity)
+                .withVariable("tableName", tableName)
+                .withVariable("keyspaceName", keyspaceName)
+                .runExpectingException(ErrorTypeMatcher.errorType(CassandraError.UNKNOWN));
+    }
 }
